make seen movies button work and improve styling for sort button

+++ backend/src/main/java/com/tmanea/backend/service/ActiveMovieService.java
@@ -71,7 +71,7 @@ public class ActiveMovieService {
             boolean stillActive = scrapedMovies.stream()
                     .anyMatch(scraped -> scraped.getTitle().equalsIgnoreCase(movie.getTitle()));
 
-            if (!stillActive) {
+            if (!stillActive && !movie.isSeen()) {
                 repository.deleteById(movie.getId());
                 log.info("MOVIE DELETED FROM DB WITH TITLE {}", movie.getTitle());
             }
